+++ python/paddle/fluid/tests/unittests/xpu/test_unsqueeze2_op_xpu.py
@@ -38,7 +38,7 @@ class TestUnsqueezeOp(XPUOpTest):
         self.init_attrs()
         self.outputs = {
             "Out": self.inputs["X"].reshape(self.new_shape),
-            "XShape": np.random.random(self.ori_shape).astype("float64")
+            "XShape": np.random.random(self.ori_shape).astype("float32")
         }
 
     def test_check_output(self):
